+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [9.12.0] - 2022-03-30
 ### Added
 - Honor video color range in VideoToolbox
 - Support for 4:2:2 and 4:4:4 pixel formats (8 and 10 bits) in VideoToolbox